fix(auth): cast $4 to bigint in register INSERT to resolve asyncpg type ambiguity
trial 值同时作 bigint 列值与 CASE WHEN 比较，asyncpg 推不出类型（integer vs bigint）。

+++ backend/src/services/apikey/users.py
@@ -69,8 +69,8 @@ async def register(email: str, password: str, *, device_id: Optional[str] = None
             """
             INSERT INTO ak_users (email, password_hash, role, balance_micro_usd,
                                   trial_micro_usd, trial_expires_at)
-            VALUES ($1, $2, $3, 0, $4,
-                    CASE WHEN $4 > 0 THEN now() + make_interval(days => $5) ELSE NULL END)
+            VALUES ($1, $2, $3, 0, $4::bigint,
+                    CASE WHEN $4::bigint > 0 THEN now() + make_interval(days => $5) ELSE NULL END)
             RETURNING *
             """,
             email, pwd_hash, role, trial, int(trial_days),
